Delete button working
Reason: Hypothesis: The argument for the make_button function was changed from a multi_item object to an int argument representing the multi_item's panel number. This changed fixed the problem because when a multi_item is passed in, all functions which use the multi_item argument or can possibly reference the object are evaluated immediately. This means that get_position in the button_operator() function's delete option is assigned at the time of the make_button's call. This is a problem because the position of the multi_item may change as more are deleted or added. Therefore, by changing the argument to an integer, the button_operator() function is not evaluated until the execute() function is called and therefore get_position is decided at the time of execution rather than button creation. This was the first and only change which first solved the problem.